refactor(occt-wasm): extract transformOps.ts (#927)
Stacked on PR #926. Extracts: composeTransform, transform, translate,
rotate, mirror, scale, generalTransform, generalTransformNonOrthogonal,
positionOnCurve, linearPattern, circularPattern, transformBatch.
multiplyMatrices4x4 helper relocated to helpers.ts.
